pipeline/configs/search-engine/requirements.txt
kafka-python
python-dotenv
qdrant-client
requests